Fixed Layout issue partially
Added 'document.close()' this fixes  component partials being treated as layouts. This also allows you to go to the updated layout. other wise the error you come up with is `Uncaught TypeError: Cannot read property 'context' of undefined`. What i cannot solve however is when the document clears because of `document.write()` it has a white flash because of the rendering.

@LukeTowers perhaps you can take a look at this https://stackoverflow.com/a/4329274 and see if it provides a possible solution?